# src/swagger.json
{
  "definitions": {},
  "info": {
    "title": "cinematev2",
    "version": "1.0.0",
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3)."
  },
  "paths": {
    "/signup": {
      "get": {
        "summary": "/signup",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/login": {
      "get": {
        "summary": "/login",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/home": {
      "get": {
        "summary": "/home",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/": {
      "get": {
        "summary": "/",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/dashboard": {
      "get": {
        "summary": "/dashboard",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/signup": {
      "post": {
        "summary": "/api/auth/signup",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "/api/auth/login",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "/api/auth/change-password",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "/api/auth/logout",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/delete": {
      "delete": {
        "summary": "/api/auth/delete",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/auth/profile/pfp": {
      "put": {
        "summary": "/api/auth/profile/pfp",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/series/{id}": {
      "get": {
        "summary": "/series/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string",
              "example": "<!DOCTYPE html>\r\n<html lang=\"en\">\r\n<head>\r\n  <meta charset=\"UTF-8\" />\r\n  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\" />\r\n  <title> Details</title>\r\n  <link rel=\"stylesheet\" href=\"/css/series-details.css\" />\r\n  <script type=\"module\" src=\"https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js\"></script>\r\n  <script nomodule src=\"https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js\"></script>\r\n</head>\r\n<body data-media-id=\"1544633\">\r\n\r\n  <!-- NAVIGATION -->\r\n  <nav>\r\n    <div class=\"logo\">\r\n      <ion-icon name=\"film-outline\"></ion-icon>\r\n      <span>Cinemate</span>\r\n    </div>\r\n\r\n    <ul>\r\n      <li><a href=\"/home\">Home</a></li>\r\n      <li><a href=\"/settings\">Settings</a></li>\r\n    </ul>\r\n\r\n    <div class=\"nav-actions\">\r\n      <a href=\"/settings/profile\"><ion-icon name=\"person-circle-outline\"></ion-icon></a>\r\n    </div>\r\n  </nav>\r\n\r\n  <!-- HERO SECTION -->\r\n  <section class=\"hero\" style=\"background-image: url('https://image.tmdb.org/t/p/originalnull');\">\r\n    <div class=\"overlay\"></div>\r\n    <div class=\"hero-content\">\r\n      <h1></h1>\r\n      <p class=\"tagline\"></p>\r\n      <div class=\"hero-buttons\">\r\n        <button class=\"btn-primary\">Add to Watchlist</button>\r\n        <button class=\"btn-favorite\" id=\"favoriteBtn\">❤ Add to Favorites</button>\r\n        <button class=\"btn-watched\" id=\"watchedBtn\">✓ Mark as Watched</button>\r\n        <button class=\"btn-secondary\" id=\"sharebtn\">Share</button>\r\n      </div>\r\n    </div>\r\n  </section>\r\n\r\n  <main>\r\n    <!-- OVERVIEW -->\r\n    <section class=\"overview\">\r\n      <div class=\"poster\">\r\n        <img src=\"https://image.tmdb.org/t/p/w500/tUbFaLY6Ys9T2wTNvPUJARi5Eri.jpg\" alt=\" Poster\">\r\n      </div>\r\n      <div class=\"info\">\r\n        <h2>Overview</h2>\r\n        <p>A widower&#39;s life is turned upside down when he begins a romance with a television personality.</p>\r\n\r\n        <div class=\"meta\">\r\n          <p><strong>Genres:</strong> Comedy, Romance</p>\r\n          <p><strong>First Air Date:</strong> </p>\r\n          <p><strong>Seasons:</strong> 0</p>\r\n          <p><strong>Episodes:</strong> 1</p>\r\n          <p><strong>Rating:</strong> ⭐ 0/10</p>\r\n          <p><strong>Status:</strong> Released</p>\r\n          <p><strong>Language:</strong> LT</p>\r\n        </div>\r\n      </div>\r\n    </section>\r\n\r\n    <!-- STAR RATING -->\r\n    <div class=\"rating-container\">\r\n      <div class=\"stars\" id=\"starRating\">\r\n        \r\n          <span data-value=\"1\">★</span>\r\n        \r\n          <span data-value=\"2\">★</span>\r\n        \r\n          <span data-value=\"3\">★</span>\r\n        \r\n          <span data-value=\"4\">★</span>\r\n        \r\n          <span data-value=\"5\">★</span>\r\n        \r\n          <span data-value=\"6\">★</span>\r\n        \r\n          <span data-value=\"7\">★</span>\r\n        \r\n          <span data-value=\"8\">★</span>\r\n        \r\n          <span data-value=\"9\">★</span>\r\n        \r\n          <span data-value=\"10\">★</span>\r\n        \r\n      </div>\r\n      <p id=\"ratingValue\">Your rating: 0 / 10</p>\r\n      <button id=\"submitRating\">Submit Rating</button>\r\n    </div>\r\n\r\n    <!-- SEASONS -->\r\n    <section class=\"seasons\">\r\n      <h2>Seasons</h2>\r\n      <div class=\"season-grid\">\r\n        \r\n          <p>No seasons available.</p>\r\n        \r\n      </div>\r\n\r\n      <!-- SEASON PAGINATION -->\r\n      <div class=\"pagination\">\r\n        \r\n      </div>\r\n\r\n      <!-- EPISODES -->\r\n      <h2>Episodes - Season 1</h2>\r\n      <div class=\"episode-grid\">\r\n        \r\n          \r\n            <div class=\"episode-card\">\r\n              <div class=\"episode-thumb\">\r\n                \r\n                  <img src=\"https://image.tmdb.org/t/p/w300/tUbFaLY6Ys9T2wTNvPUJARi5Eri.jpg\" alt=\"Laimingos žvaigždės\">\r\n                \r\n                    <button class=\"btn-watched-ep\" \r\n                            data-season=\"\" \r\n                            data-episode=\"1\">\r\n                      ✓ Watched\r\n                    </button>\r\n\r\n              </div>\r\n              <div class=\"episode-info\">\r\n                <div class=\"episode-left\">\r\n                  <span class=\"episode-number\">SE1</span>\r\n                  <span class=\"episode-title\">Laimingos žvaigždės</span>\r\n                  <p class=\"episode-overview\">\r\n                    A widower&#39;s life is turned upside down when he begins a romance with a television personality.\r\n                  </p>\r\n                </div>\r\n                <div class=\"episode-right\">\r\n                  <span class=\"episode-airdate\">2025-10-31</span>\r\n                </div>\r\n              </div>\r\n            </div>\r\n          \r\n        \r\n      </div>\r\n\r\n      <!-- EPISODE PAGINATION -->\r\n      <div class=\"pagination\">\r\n        \r\n          <a class=\"active\" \r\n             href=\"?season=1&episodePage=1\">\r\n            1\r\n          </a>\r\n        \r\n      </div>\r\n    </section>\r\n\r\n    <!-- CAST -->\r\n    <section class=\"cast\">\r\n      <h2>Cast</h2>\r\n      <div class=\"cast-grid\">\r\n        \r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/bqiuVc4Hy0mZdXEkngWhheTA4rj.jpg\" alt=\"Justinas Jankevičius\">\r\n              <p class=\"actor-name\">Justinas Jankevičius</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/otm4XV2huIENXjxCW6IXMwOxcz6.jpg\" alt=\"Ineta Stasiulytė\">\r\n              <p class=\"actor-name\">Ineta Stasiulytė</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/A8smSE4bjk713DB6BfiqKwnNxFe.jpg\" alt=\"Valentinas Krulikovskis\">\r\n              <p class=\"actor-name\">Valentinas Krulikovskis</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/6L8se4vVD3zbMigY2ybluXA0wsY.jpg\" alt=\"Mindaugas Capas\">\r\n              <p class=\"actor-name\">Mindaugas Capas</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/7vDG3AMYLZ0GBk5OqSCFDtvb0cM.jpg\" alt=\"Larisa Kalpokaitė\">\r\n              <p class=\"actor-name\">Larisa Kalpokaitė</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/r6Gdi2P7D6cPHcgTvvO4NSzvb9I.jpg\" alt=\"Liubomiras Laucevicius\">\r\n              <p class=\"actor-name\">Liubomiras Laucevicius</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185\" alt=\"Rolandas Mackevičius\">\r\n              <p class=\"actor-name\">Rolandas Mackevičius</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185\" alt=\"Teodoras Lipčius\">\r\n              <p class=\"actor-name\">Teodoras Lipčius</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/82VrI3eus5Zma08TVwTFUnCDKWC.jpg\" alt=\"Sandra Daukšaitė-Petrulėnė\">\r\n              <p class=\"actor-name\">Sandra Daukšaitė-Petrulėnė</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n            <div class=\"cast-card\">\r\n              <img src=\"https://image.tmdb.org/t/p/w185/9E2P758vLUbCj1mIoYsXXMijJ0k.jpg\" alt=\"Aldona Vilutytė\">\r\n              <p class=\"actor-name\">Aldona Vilutytė</p>\r\n              <small></small>\r\n            </div>\r\n          \r\n        \r\n      </div>\r\n    </section>\r\n\r\n    <!-- COMMENTS -->\r\n    <section class=\"comments\">\r\n      <h2>Comments</h2>\r\n      <div class=\"comments-container\"></div>\r\n      <div class=\"comment-form\">\r\n        <textarea id=\"newComment\" placeholder=\"Write a comment...\"></textarea>\r\n        <button id=\"addCommentBtn\">Send</button>\r\n      </div>\r\n    </section>\r\n  </main>\r\n\r\n  <script src=\"/js/theme.js\"></script>\r\n  <script src=\"/js/series-details.js\"></script>\r\n</body>\r\n</html>\r\n"
            }
          },
          "304": {}
        },
        "tags": [],
        "produces": [
          "text/html"
        ]
      }
    },
    "/series/{id}/comments": {
      "get": {
        "summary": "/series/{id}/comments",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "304": {}
        },
        "tags": []
      },
      "post": {
        "summary": "/series/{id}/comments",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/series/{id}/rate": {
      "post": {
        "summary": "/series/{id}/rate",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/series/{id}/rating": {
      "get": {
        "summary": "/series/{id}/rating",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "304": {}
        },
        "tags": []
      }
    },
    "/series/{id}/watchlist/check": {
      "get": {
        "summary": "/series/{id}/watchlist/check",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inWatchlist": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "304": {}
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/watchlist/toggle": {
      "post": {
        "summary": "/series/{id}/watchlist/toggle",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inWatchlist": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/favorites/check": {
      "get": {
        "summary": "/series/{id}/favorites/check",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inFavorites": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "304": {}
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/favorites/toggle": {
      "post": {
        "summary": "/series/{id}/favorites/toggle",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inFavorites": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/watched/check": {
      "get": {
        "summary": "/series/{id}/watched/check",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "isWatched": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "304": {}
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/watched/toggle": {
      "post": {
        "summary": "/series/{id}/watched/toggle",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "isWatched": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        },
        "tags": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/series/{id}/progress": {
      "get": {
        "summary": "/series/{id}/progress",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "304": {}
        },
        "tags": []
      }
    },
    "/settings": {
      "get": {
        "summary": "/settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/settings/password": {
      "get": {
        "summary": "/settings/password",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/settings/profile": {
      "get": {
        "summary": "/settings/profile",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/media": {
      "get": {
        "summary": "/api/media",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/media/search": {
      "get": {
        "summary": "/api/media/search",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api/media/sync": {
      "get": {
        "summary": "/api/media/sync",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/user_media/watchlist": {
      "get": {
        "summary": "/user_media/watchlist",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/user_media/watchlist/{mediaId}": {
      "delete": {
        "summary": "/user_media/watchlist/{mediaId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/user_media/finished": {
      "get": {
        "summary": "/user_media/finished",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/user_media/watching": {
      "get": {
        "summary": "/user_media/watching",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/admin": {
      "get": {
        "summary": "/admin",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/admin/make_admin": {
      "post": {
        "summary": "/admin/make_admin",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/api-docs/json": {
      "get": {
        "summary": "/api-docs/json",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    }
  },
  "swagger": "2.0",
  "tags": []
}